requirement.txt
pandas

plotly-express

streamlit
